FacialSCDnet+/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
alembic=1.13.1=pyhd8ed1ab_1
anyio=4.2.0=pyhd8ed1ab_0
bcrypt=4.1.2=pypi_0
blas=1.0=mkl
blinker=1.7.0=pyhd8ed1ab_0
bottleneck=1.3.7=pypi_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=pypi_0
bzip2=1.0.8=hd590300_5
ca-certificates=2024.3.11=h06a4308_0
certifi=2024.2.2=pypi_0
cffi=1.16.0=pypi_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
click=8.1.7=unix_pyh707e725_0
cloudpickle=3.0.0=pyhd8ed1ab_0
colorlog=5.0.1=pypi_0
configparser=5.3.0=pyhd8ed1ab_0
contourpy=1.2.0=pypi_0
cryptography=42.0.3=pypi_0
cuda-cudart=11.7.99=0
cuda-cupti=11.7.101=0
cuda-libraries=11.7.1=0
cuda-nvrtc=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-runtime=11.7.1=0
cycler=0.11.0=pyhd3eb1b0_0
databricks-cli=0.18.0=pyhd8ed1ab_0
docker-py=7.0.0=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
exceptiongroup=1.2.0=pyhd8ed1ab_2
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=pypi_0
flask=3.0.2=pyhd8ed1ab_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
gitdb=4.0.11=pyhd8ed1ab_0
gitpython=3.1.42=pyhd8ed1ab_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=pypi_0
gnutls=3.6.15=he1e5248_0
gorilla=0.4.0=pyhd8ed1ab_0
greenlet=3.0.1=pypi_0
gunicorn=21.2.0=pypi_0
h11=0.14.0=pyhd8ed1ab_0
h2=4.1.0=pyhd8ed1ab_0
hpack=4.0.0=pyh9f0ad1d_0
httpcore=1.0.3=pyhd8ed1ab_0
httpx=0.26.0=pyhd8ed1ab_0
hyperframe=6.0.1=pyhd8ed1ab_0
idna=3.6=pyhd8ed1ab_0
importlib-metadata=7.0.1=pyha770c72_0
importlib_resources=6.1.1=pyhd8ed1ab_0
intel-openmp=2023.1.0=hdb19cb5_46306
itsdangerous=2.1.2=pyhd8ed1ab_0
jinja2=3.1.3=pyhd8ed1ab_0
joblib=1.3.2=pyhd8ed1ab_0
jpeg=9e=h5eee18b_1
kiwisolver=1.4.4=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.40=h41732ed_0
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libcublas=11.10.3.66=0
libcufft=10.7.2.124=h4fbf590_0
libcufile=1.8.1.2=0
libcurand=10.3.4.107=0
libcusolver=11.4.0.1=0
libcusparse=11.7.4.91=0
libdeflate=1.8=h7f8727e_5
libexpat=2.5.0=hcb278e6_1
libffi=3.4.2=h7f98852_5
libgcc-ng=13.2.0=h807b86a_5
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=13.2.0=h807b86a_5
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libnpp=11.7.4.75=0
libnsl=2.0.1=hd590300_0
libnvjpeg=11.8.0.2=0
libopenblas=0.3.21=h043d6bf_0
libpng=1.6.39=h5eee18b_0
libprotobuf=3.20.3=he621ea3_0
libsodium=1.0.18=h36c2ea0_1
libsqlite=3.45.1=h2797004_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.4.0=hecacb30_2
libunistring=0.9.10=h27cfd23_0
libuuid=2.38.1=h0b41bf4_0
libwebp-base=1.3.2=h5eee18b_0
libxcrypt=4.4.36=hd590300_1
libzlib=1.2.13=hd590300_5
lz4-c=1.9.4=h6a678d5_0
mako=1.3.2=pyhd8ed1ab_0
markupsafe=2.1.5=pypi_0
matplotlib=3.8.4=pypi_0
matplotlib-base=3.8.4=py311ha02d727_0
mkl=2023.1.0=h213fc3f_46344
mkl-fft=1.3.8=pypi_0
mkl-random=1.2.4=pypi_0
mkl-service=2.4.0=pypi_0
mkl_fft=1.3.8=py311h5eee18b_0
mkl_random=1.2.4=py311hdb19cb5_0
mlflow=1.2.0=py_1
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=pypi_0
munkres=1.1.4=py_0
ncurses=6.4=h59595ed_2
nettle=3.7.3=hbbd107a_1
networkx=3.1=pypi_0
numexpr=2.8.7=pypi_0
numpy=1.26.3=pypi_0
numpy-base=1.26.3=py311hf175353_0
oauthlib=3.2.2=pyhd8ed1ab_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.2.1=hd590300_1
optuna=3.5.0=pyhd8ed1ab_0
packaging=23.2=pyhd8ed1ab_0
pandas=2.1.4=pypi_0
paramiko=3.4.0=pyhd8ed1ab_0
pillow=10.2.0=pypi_0
pip=24.0=pyhd8ed1ab_0
protobuf=3.20.3=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pyjwt=2.8.0=pyhd8ed1ab_1
pynacl=1.5.0=pypi_0
pyopenssl=24.0.0=pyhd8ed1ab_0
pyparsing=3.0.9=pypi_0
pysocks=1.7.1=pyha2e5f31_6
python=3.11.8=hab00c5b_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-telegram-bot=20.8=pyhd8ed1ab_0
python-tzdata=2024.1=pyhd8ed1ab_0
python_abi=3.11=4_cp311
pytorch=2.0.1=py3.11_cuda11.7_cudnn8.5.0_0
pytorch-cuda=11.7=h778d358_5
pytorch-mutex=1.0=cuda
pytz=2024.1=pyhd8ed1ab_0
pywin32-on-windows=0.1.0=pyh1179c8e_3
pyyaml=6.0.1=pypi_0
querystring_parser=1.2.4=py_0
readline=8.2=h8228510_1
requests=2.31.0=pyhd8ed1ab_0
scikit-learn=1.2.2=pypi_0
scipy=1.11.4=pypi_0
seaborn=0.12.2=pypi_0
setuptools=69.0.3=pyhd8ed1ab_0
simplejson=3.19.2=pypi_0
six=1.16.0=pyh6c4a22f_0
smmap=5.0.0=pyhd8ed1ab_0
sniffio=1.3.0=pyhd8ed1ab_0
sqlalchemy=2.0.27=pypi_0
sqlparse=0.4.4=pyhd8ed1ab_0
sympy=1.12=pypi_0
tabulate=0.9.0=pyhd8ed1ab_1
tbb=2021.8.0=hdb19cb5_0
threadpoolctl=3.3.0=pyhc1e730c_0
tk=8.6.13=noxft_h4845f30_101
torch=2.0.1=pypi_0
torchtriton=2.0.0=py311
torchvision=0.15.2=pypi_0
tqdm=4.65.0=pypi_0
triton=2.0.0=pypi_0
typing-extensions=4.9.0=hd8ed1ab_0
typing_extensions=4.9.0=pyha770c72_0
tzdata=2024a=h0c530f3_0
urllib3=1.26.15=pyhd8ed1ab_0
websocket-client=1.7.0=pyhd8ed1ab_0
werkzeug=3.0.1=pyhd8ed1ab_0
wheel=0.42.0=pyhd8ed1ab_0
xz=5.2.6=h166bdaf_0
yaml=0.2.5=h7f98852_2
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=hd590300_5
zstd=1.5.2=ha4553b6_0
